feat: [임보경] 앱에서 직접 구매 버튼 기능 추가

+++ ztz_process/src/main/java/kr/eddi/ztz_process/controller/order/CartController.java
@@ -53,7 +53,11 @@ public class CartController {
     @DeleteMapping(path = "/{itemNo}", headers = "Token")
     public String deleteItem (@PathVariable("itemNo") Long itemNo, @RequestHeader("Token") String token) {
         log.info("선택된 단일 항목 삭()" + itemNo + " : " + token );
-        service.deleteCartItem(itemNo, token);
+        if(itemNo == 1000001) {
+            log.info("direct 구매 ");
+        } else {
+            service.deleteCartItem(itemNo, token);
+        }
         return "1";
     }
